Seemingly fixed bug with starting app from notification. Should all be ready now to plug in real notifications from backend

+++ app/src/main/java/com/joshrap/liteweight/helpers/InputHelper.java
@@ -79,7 +79,7 @@ public class InputHelper {
             retVal = "Weight cannot be empty";
         } else {
             try {
-                int weight = Integer.parseInt(aWeight);
+                double weight = Double.parseDouble(aWeight);
                 if (weight < 0 || weight > Variables.MAX_WEIGHT) {
                     retVal = String.format("Enter value between 1-%s.", Variables.MAX_WEIGHT);
                 }
